requirements.txt
Django==4.1.3
django-geojson==4.0.0
django-leaflet==0.28.3
psycopg2==2.9.5
psycopg2-binary==2.9.5
